pkg/seq/parser/match: renames the package from `parse` to better distinguish from parent package
